renamed fixed-number-helper
