+++ src/helpers/fixednumber-helper.ts
